# docs/swagger.yaml
basePath: /api/v1
definitions:
  handlers.CreateBookingRequest:
    properties:
      booking_name:
        type: string
      check_in_date:
        type: string
      check_out_date:
        type: string
      room_num:
        type: integer
    required:
    - booking_name
    - check_in_date
    - check_out_date
    - room_num
    type: object
...
  handlers.GetAvailableRoomsRequest:
    properties:
      check_in_date:
        type: string
      check_out_date:
        type: string
    required:
    - check_in_date
    - check_out_date
    type: object
  handlers.GetBookingsByDateRangeRequest:
    properties:
      end_date:
        type: string
      start_date:
        type: string
    required:
    - end_date
    - start_date
    type: object
...
  handlers.UpdateBookingRequest:
    properties:
      check_in_date:
        type: string
      check_out_date:
        type: string
    required:
    - check_in_date
    - check_out_date
    type: object
  models.Booking:
    properties:
      booking_date:
        type: string
      booking_id:
        type: integer
      booking_name:
        type: string
      check_in_date:
        type: string
      check_out_date:
        type: string
      receipt:
        $ref: '#/definitions/models.Receipt'
      room:
        $ref: '#/definitions/models.Room'
      room_num:
        type: integer
      statuses:
        items:
          $ref: '#/definitions/models.RoomStatus'
        type: array
      total_price:
        type: integer
    required:
    - booking_name
    - check_in_date
    - check_out_date
    - room_num
    type: object
  models.Receipt:
    properties:
      amount:
        type: integer
      booking_id:
        type: integer
      issue_date:
        type: string
      payment_date:
        type: string
      payment_method:
        enum:
        - Credit
        - Debit
        - Bank
        - Transfer
        type: string
      receipt_id:
        type: integer
    required:
    - amount
    - payment_method
    type: object
  models.Room:
    properties:
      bookings:
        items:
          $ref: '#/definitions/models.Booking'
        type: array
      room_num:
        type: integer
      room_type:
        $ref: '#/definitions/models.RoomType'
      statuses:
        items:
          $ref: '#/definitions/models.RoomStatus'
        type: array
      type_id:
        type: integer
    type: object
  models.RoomFacility:
    properties:
      fac_id:
        type: integer
      type_id:
        type: integer
    type: object
  models.RoomStatus:
    properties:
      booking:
        $ref: '#/definitions/models.Booking'
      booking_id:
        type: integer
      calendar:
        type: string
      room_num:
        type: integer
      status:
        enum:
        - Available
        - Occupied
        type: string
    required:
    - status
    type: object
  models.RoomType:
    properties:
      area:
        type: integer
      description:
        type: string
      name:
        type: string
      noOfGuest:
        type: integer
      price_per_night:
        type: integer
      room_facilities:
        items:
          $ref: '#/definitions/models.RoomFacility'
        type: array
      rooms:
        items:
          $ref: '#/definitions/models.Room'
        type: array
      type_id:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for hotel booking system with transaction management and concurrency
    control
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Hotel Booking System API
  version: "1.0"
paths:
  /bookings:
    post:
      consumes:
      - application/json
      description: Create a new booking for a room
      parameters:
      - description: Booking details
        in: body
        name: booking
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateBookingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Booking'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new booking
      tags:
      - bookings
  /bookings/{id}:
    delete:
      description: Cancel a booking by ID
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Cancel a booking
      tags:
      - bookings
    get:
      description: Get a booking by ID
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Booking'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a booking
      tags:
      - bookings
    put:
      consumes:
      - application/json
      description: Update a booking's dates
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: integer
      - description: New booking dates
        in: body
        name: booking
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateBookingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a booking
      tags:
      - bookings
  /bookings/by-date:
    post:
      consumes:
      - application/json
      description: Get all bookings within a date range
      parameters:
      - description: Date range
        in: body
        name: dates
        required: true
        schema:
          $ref: '#/definitions/handlers.GetBookingsByDateRangeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Booking'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get bookings by date range
      tags:
      - bookings
...
  /rooms/available:
    post:
      consumes:
      - application/json
      description: Get available rooms for a date range
      parameters:
      - description: Date range
        in: body
        name: dates
        required: true
        schema:
          $ref: '#/definitions/handlers.GetAvailableRoomsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Room'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get available rooms
      tags:
      - rooms
...
swagger: "2.0"